requirements.txt
streamlit
google-generativeai
reportlab
firebase-admin